chore: include shared package in docker builds
