+++ backend/docker-compose.override.yml
@@ -1,9 +1,10 @@
 services:
   backend:
     build:
-      dockerfile: Dockerfile.dev
+      dockerfile: backend/Dockerfile.dev
     volumes:
       - .:/app
+      - ../shared:/shared
       - /app/node_modules
     env_file:
       - .env.development
