+++ backend/package.json
@@ -108,6 +108,9 @@
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
     },
+    "moduleNameMapper": {
+      "^@rflp/shared(.*)$": "<rootDir>/../../shared/src$1"
+    },
     "collectCoverageFrom": [
       "**/*.(t|j)s"
     ],
