+++ frontend/docker-compose.yml
@@ -3,8 +3,8 @@ services:
     container_name: frontend
     image: rflandscaperpro/frontend
     build:
-      context: .
-      dockerfile: Dockerfile
+      context: ..
+      dockerfile: frontend/Dockerfile
     ports:
       - "4000:4000"
     networks:
